+++ changelog.md
@@ -31,6 +31,7 @@
 - Unit test for InvoiceItemWrapper
 - Rename InvoiceItems to InvoiceItemsTable
 - add test for InvoiceItemsTable
+- add error boundary
 # 2022 04 01
 - add unit testing for Avatar
 - Add unit testing for ProfileForm, add more test cases